Move logic out of v1util into other packages, leave aliases for back-compat.
