+++ pkg/v1/internal/and/and_closer.go
@@ -1,4 +1,4 @@
-// Copyright 2018 Google LLC All Rights Reserved.
+// Copyright 2020 Google LLC All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,36 +12,36 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1util
+package and
 
 import (
 	"io"
 )
 
-// readAndCloser implements io.ReadCloser by reading from a particular io.Reader
+// ReadCloser implements io.ReadCloser by reading from a particular io.Reader
 // and then calling the provided "Close()" method.
-type readAndCloser struct {
+type ReadCloser struct {
 	io.Reader
 	CloseFunc func() error
 }
 
-var _ io.ReadCloser = (*readAndCloser)(nil)
+var _ io.ReadCloser = (*ReadCloser)(nil)
 
 // Close implements io.ReadCloser
-func (rac *readAndCloser) Close() error {
+func (rac *ReadCloser) Close() error {
 	return rac.CloseFunc()
 }
 
-// writeAndCloser implements io.WriteCloser by reading from a particular io.Writer
+// WriteCloser implements io.WriteCloser by reading from a particular io.Writer
 // and then calling the provided "Close()" method.
-type writeAndCloser struct {
+type WriteCloser struct {
 	io.Writer
 	CloseFunc func() error
 }
 
-var _ io.WriteCloser = (*writeAndCloser)(nil)
+var _ io.WriteCloser = (*WriteCloser)(nil)
 
 // Close implements io.WriteCloser
-func (wac *writeAndCloser) Close() error {
+func (wac *WriteCloser) Close() error {
 	return wac.CloseFunc()
 }
